huge WIP refactor of GameImporter
1.  Introducing a robust ImportQueueItem to track an import and follow it through the system
2.  Refactored code to extract behaviors out of GameImporter so it's not a god class
3.  Pre compute as much as possible for an ImportQueueItem, such as which systems are valid.
4.  Try to simplify the general logic to 3 steps - compute, move, and import into the DB

Remaining work

1.  Tons of code clean up
2.  Proper handling of conflicts
3.  Proper handling of CDROMs (need to sort to handle M3U -> Cue -> Bin)
4.  Proper handling of directories
5.  Unit tests where applicable
6.  Ensure validity of imports that flow through this new system.
7.  UI for Conflict handling
8.  Validate that archives work properly.
9.  Validate and fix Artwork
