+++ PVLibrary/Tests/PVLibraryTests/PVLibraryTests.swift
@@ -14,7 +14,7 @@ class PVLibraryTests: XCTestCase {
         super.setUp()
         // Put setup code here. This method is called before the invocation of each test method in the class.
         try! FileManager.default.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: true)
-        try! RomDatabase.initDefaultDatabase()
+//        try! RomDatabase.initDefaultDatabase()
     }
 
     override func tearDown() {
